# api-specification-swagger.yaml
openapi: 3.0.0
info:
  title: Healenium API
  description: API for Healenium, a self-healing testing tool that helps to maintain web tests by automatically healing broken locators.
  version: 1.0.0
servers:
  - url: /healenium
    description: Base URL for all API endpoints
paths:
...
  /report/{uid}:
    get:
      summary: Get Report by ID
      description: Get report by ID
      tags:
        - Report Management
      parameters:
        - name: uid
          in: path
          description: Report unique ID
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            text/html:
              schema:
                type: string
...
  /report/init/{uid}:
    post:
      summary: Initialize Report by ID
      description: Initialize a report with specific ID
      tags:
        - Report Management
      parameters:
        - name: uid
          in: path
          description: Report unique ID
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            text/plain:
              schema:
                type: string
  /report/build:
    post:
      summary: Build Report
      description: Build report
      tags:
        - Report Management
      parameters:
        - name: sessionKey
          in: header
          description: Session key for report association
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            text/plain:
              schema:
                type: string
...
  /report/data/{uid}:
    get:
      summary: Get Report Data
      description: Get report data by ID
      tags:
        - Report Management
      parameters:
        - name: uid
          in: path
          description: Report unique ID
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecordDto'
        '400':
          description: Bad Request
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
    patch:
      summary: Edit Report
      description: Edit report data
      tags:
        - Report Management
      parameters:
        - name: uid
          in: path
          description: Report unique ID
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportDto'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecordDto'
...
components:
  schemas:
...
    ReportDto:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        time:
          type: string
    RecordDto:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        time:
          type: string
        data:
          type: array
          items:
            $ref: '#/components/schemas/ReportRecord'
        reportLinks:
          type: array
          items:
            $ref: '#/components/schemas/ReportLinkDto'
    ReportRecord:
      type: object
      properties:
        declaringClass:
          type: string
        screenShotPath:
          type: string
        failedLocatorValue:
          type: string
        failedLocatorType:
          type: string
        healedLocatorValue:
          type: string
        healedLocatorType:
          type: string
        score:
          type: string
        successHealing:
          type: boolean
        healingResultId:
          type: integer
          format: int32
    ReportLinkDto:
      type: object
      properties:
        name:
          type: string
        url:
          type: string
...